Changed S3MessageMarshaller to inner S3Message.Marshaller class
